move load_json_file, calculate_taxable_value and calculate_total_assessed_value function to their own class as a method
